+++ packages/components/src/popover/index.js
@@ -1,75 +1,19 @@
 /**
  * External dependencies
  */
-import {
-	useCallback,
-	useEffect,
-	useLayoutEffect,
-	useRef,
-	useState,
-} from '@wordpress/element';
+import { useLayoutEffect, useRef, useState } from '@wordpress/element';
 import classnames from 'classnames';
-import { first, forEach, some } from 'lodash';
 
 /**
  * Internal dependencies
  */
+import { useBlur } from '@crowdsignal/hooks';
 import { getPopoverOffset } from './util';
 
 /**
  * Style dependencies
  */
 import { Wrapper } from './styles';
-
-const getDocumentRoots = ( elements ) =>
-	elements
-		.map( ( { current } ) => {
-			return current ? current.ownerDocument : null;
-		} )
-		.filter( ( ownerDocument ) => !! ownerDocument )
-		.reduce(
-			( documents, ownerDocument ) => [
-				...documents,
-				...( documents.includes( ownerDocument )
-					? []
-					: [ ownerDocument ] ),
-			],
-			[]
-		);
-
-export const useOnOutsideClick = ( onOutsideClick, elements = [] ) => {
-	const handleClick = useCallback(
-		( event ) => {
-			const target = first( event.composedPath() );
-
-			if (
-				some(
-					elements,
-					( element ) =>
-						! element.current || element.current.contains( target )
-				)
-			) {
-				return;
-			}
-
-			onOutsideClick();
-		},
-		[ onOutsideClick, ...elements ]
-	);
-
-	useEffect( () => {
-		const documentRoots = getDocumentRoots( elements );
-
-		forEach( documentRoots, ( root ) =>
-			root.addEventListener( 'mousedown', handleClick )
-		);
-
-		return () =>
-			forEach( documentRoots, ( root ) =>
-				root.removeEventListener( 'mousedown', handleClick )
-			);
-	}, [ handleClick, ...elements ] );
-};
 
 const Popover = ( {
 	children,
@@ -87,9 +31,9 @@ const Popover = ( {
 		setOffset(
 			getPopoverOffset( position, popover.current, context.current )
 		);
-	}, [ context.current, popover.current, position ] );
+	}, [ context.current, isVisible, popover.current, position ] );
 
-	useOnOutsideClick( onClose, [ popover, context ] );
+	useBlur( onClose, [ popover, context ] );
 
 	const classes = classnames( 'popover', className );
