+++ packages/hooks/package.json
@@ -1,0 +1,23 @@
+{
+  "name": "@crowdsignal/hooks",
+  "version": "0.1.0",
+  "description": "Crowdsignal's component hooks.",
+  "author": "Automattic Inc.",
+  "license": "GPL-2.0-or-later",
+  "keywords": [
+    "crowdsignal"
+  ],
+  "homepage": "https://github.com/Automattic/crowdsignal-ui/tree/HEAD/packages/hooks/README.md",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/Automattic/crowdsignal-ui.git"
+  },
+  "bugs": {
+    "url": "https://github.com/Automattic/crowdsignal-ui/issues"
+  },
+  "main": "src/index.js",
+  "dependencies": {
+    "@wordpress/element": "^3.1.1",
+    "lodash": "^4.17.21"
+  }
+}
